+++ Plugins/Extensions/EpgToXml/__init__.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2026 Arcardy
 """EpgToXml Enigma2 plugin package."""
 
 PLUGIN_NAME = "EpgToXml"
-VERSION = "0.5.2"
+VERSION = "0.6.0"
 
 
 def _(text):